Rework the netlink IP filtering yet again.  Now we very carefully handle all v4_mapped and v4_compat addresses, and we query netlink with the address family of the local address.
I read the boost::asio source to see how it handled address family comparisons,
and I think this code is correct and will work.  The tricky thing is that
comparisons check the _type member first and don't even examine v4_mapped
or v4_compat -- it's a value rather than semantic equality test.

Also, another pitfall is that a ip::address that has a v6 but valid v4_mapped
or v4_compat address will throw on attempts to convert it to a v4 -- it also
checks the _type member first.  So it's necessary to convert to v6 and then
to v4.
* * *
Amend.